Moved `XorBlendOp` to original file and deleted new file
