--- Pinta.Core/Effects/UserBlendOps/XorBlendOp.cs
@@ -1,139 +1,0 @@
-using System;
-
-namespace Pinta.Core;
-
-public sealed partial class UserBlendOps
-{
-	[Serializable]
-	public sealed class XorBlendOp : UserBlendOp
-	{
-		public static string StaticName => "Xor";
-
-		public override ColorBgra Apply (in ColorBgra lhs, in ColorBgra rhs)
-		{
-			return ApplyStatic (lhs, rhs);
-		}
-
-		public override void Apply (Span<ColorBgra> dst, ReadOnlySpan<ColorBgra> src)
-		{
-			for (int i = 0; i < src.Length; ++i)
-				dst[i] = ApplyStatic (dst[i], src[i]);
-		}
-		public override void Apply (Span<ColorBgra> dst, ReadOnlySpan<ColorBgra> lhs, ReadOnlySpan<ColorBgra> rhs)
-		{
-			for (int i = 0; i < dst.Length; ++i)
-				dst[i] = ApplyStatic (lhs[i], rhs[i]);
-		}
-		public static ColorBgra ApplyStatic (in ColorBgra lhs, in ColorBgra rhs)
-		{
-			int y;
-			{
-				y = lhs.A * (255 - rhs.A) + 0x80;
-				y = ((y >> 8) + y) >> 8;
-			};
-			int totalA = y + rhs.A;
-			uint ret;
-			if (totalA == 0) {
-				ret = 0;
-			} else {
-				int fB = lhs.B ^ rhs.B;
-				int fG = lhs.G ^ rhs.G;
-				int fR = lhs.R ^ rhs.R;
-				int x;
-				{
-					x = lhs.A * rhs.A + 0x80;
-					x = ((x >> 8) + x) >> 8;
-				}
-				int z = rhs.A - x;
-				int masIndex = totalA * 3;
-				uint taM = masTable[masIndex];
-				uint taA = masTable[masIndex + 1];
-				uint taS = masTable[masIndex + 2];
-				uint b = (uint) (((((long) ((lhs.B * y) + (rhs.B * z) + (fB * x))) * taM) + taA) >> (int) taS);
-				uint g = (uint) (((((long) ((lhs.G * y) + (rhs.G * z) + (fG * x))) * taM) + taA) >> (int) taS);
-				uint r = (uint) (((((long) ((lhs.R * y) + (rhs.R * z) + (fR * x))) * taM) + taA) >> (int) taS);
-				int a;
-				{
-					a = lhs.A * (255 - rhs.A) + 0x80;
-					a = ((a >> 8) + a) >> 8;
-					a += rhs.A;
-				}
-				ret = b + (g << 8) + (r << 16) + ((uint) a << 24);
-			};
-			return ColorBgra.FromUInt32 (ret);
-		}
-		public override UserBlendOp CreateWithOpacity (int opacity) => new XorBlendOpWithOpacity (opacity);
-		private sealed class XorBlendOpWithOpacity : UserBlendOp
-		{
-			private readonly int opacity;
-			private byte ApplyOpacity (byte a)
-			{
-				int r;
-				{
-					r = a;
-					r = r * opacity + 0x80;
-					r = ((r >> 8) + r) >> 8;
-				}
-				return (byte) r;
-			}
-			public static string StaticName => "UserBlendOps." + "Xor" + "BlendOp.Name";
-			public override ColorBgra Apply (in ColorBgra lhs, in ColorBgra rhs)
-			{
-				int lhsA = lhs.A;
-				int rhsA = ApplyOpacity (rhs.A);
-				int y;
-				{
-					y = lhsA * (255 - rhsA) + 0x80;
-					y = ((y >> 8) + y) >> 8;
-				};
-				int totalA = y + rhsA;
-				uint ret;
-				if (totalA == 0) {
-					ret = 0;
-				} else {
-					int fB = lhs.B ^ rhs.B;
-					int fG = lhs.G ^ rhs.G;
-					int fR = lhs.R ^ rhs.R;
-					int x;
-					{
-						x = lhsA * rhsA + 0x80;
-						x = ((x >> 8) + x) >> 8;
-					}
-					int z = rhsA - x;
-					int masIndex = totalA * 3;
-					uint taM = masTable[masIndex];
-					uint taA = masTable[masIndex + 1];
-					uint taS = masTable[masIndex + 2];
-					uint b = (uint) (((((long) ((lhs.B * y) + (rhs.B * z) + (fB * x))) * taM) + taA) >> (int) taS);
-					uint g = (uint) (((((long) ((lhs.G * y) + (rhs.G * z) + (fG * x))) * taM) + taA) >> (int) taS);
-					uint r = (uint) (((((long) ((lhs.R * y) + (rhs.R * z) + (fR * x))) * taM) + taA) >> (int) taS);
-					int a;
-					{
-						a = lhsA * (255 - rhsA) + 0x80;
-						a = ((a >> 8) + a) >> 8;
-						a += rhsA;
-					}
-					ret = b + (g << 8) + (r << 16) + ((uint) a << 24);
-				}
-				return ColorBgra.FromUInt32 (ret);
-			}
-			public override void Apply (Span<ColorBgra> dst, ReadOnlySpan<ColorBgra> src)
-			{
-				for (int i = 0; i < src.Length; ++i)
-					dst[i] = ApplyStatic (dst[i], src[i]);
-			}
-			public override void Apply (Span<ColorBgra> dst, ReadOnlySpan<ColorBgra> lhs, ReadOnlySpan<ColorBgra> rhs)
-			{
-				for (int i = 0; i < dst.Length; ++i)
-					dst[i] = ApplyStatic (lhs[i], rhs[i]);
-			}
-			public XorBlendOpWithOpacity (int opacity)
-			{
-				if (opacity < 0 || opacity > 255)
-					throw new ArgumentOutOfRangeException (nameof (opacity));
-
-				this.opacity = opacity;
-			}
-		}
-	}
-}
